Mathematical_solution/Backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.6.1
manim==0.18.0
manim-physics==0.3.0
numpy==1.26.3
scipy==1.11.4
pillow==10.1.0
opencv-python==4.8.1.78
crewai==0.28.8
google-generativeai==0.3.2
tiktoken==0.5.2
langchain==0.1.11
requests==2.31.0
aiofiles==23.2.1
watchdog==3.0.0
loguru==0.7.2
pytest==8.0.2
httpx==0.26.0
python-magic==0.4.27
PyPDF2==3.0.1
pdf2image==1.16.3